// src/state/types.ts
export type SerializedBigNumber = string
// ...
export interface PresaleState {
  presaleStartTimestamp?: SerializedBigNumber
  presaleEndTimestamp?: SerializedBigNumber
  hardCapEthAmount?: SerializedBigNumber
  totalDepositedEthBalance?: SerializedBigNumber
  minimumDepositEthAmount?: SerializedBigNumber
  maximumDepositEthAmount?: SerializedBigNumber
  userDataLoaded?: boolean
  userData?: {
    balance: SerializedBigNumber
    deposits: SerializedBigNumber
  }
}